added ability to specific hardware endianness
While column 0 is meant to indicate the left most column, the hardware layout of the matrix might not
make the left most column the MSB in the shift register chain. This might happen if you make a mistake in your
hardware (oops). By setting the column endian value of the matrix,the high level software can still reference
column 0 to mean the left most column, and the bit layout sent the shift registers will be adjusted according
to where column 0 physically is in the hardware.